changed format of price to int, redirectUrl for payment is taken from…

+++ src/main/kotlin/fi/metatavu/ikioma/persistence/dao/PrescriptionRenewalDAO.kt
@@ -37,7 +37,7 @@ class PrescriptionRenewalDAO : AbstractDAO<PrescriptionRenewal>() {
         paymentStatus: PaymentStatus,
         paymentUrl: String,
         stamp: UUID,
-        price: Long,
+        price: Int,
         checkoutAccount: Int,
         creatorId: UUID,
         lastModifierId: UUID
